feat: add localize date with billing information
