+++ package.json
@@ -50,6 +50,7 @@
     "lodash.camelcase": "^4.3.0",
     "lodash.snakecase": "^4.1.1",
     "lodash.without": "^4.4.0",
+    "moment-timezone": "^0.5.43",
     "prop-types": "^15.8.1",
     "react": "^16.14.0",
     "react-dom": "^16.14.0",
