+++ src/subscription/checkout/monthly-billing-notification/MonthlyBillingNotification.jsx
@@ -1,33 +1,35 @@
-import React from 'react';
+import React, { useContext } from 'react';
 import { useSelector } from 'react-redux';
-import { useIntl, defineMessages } from '@edx/frontend-platform/i18n';
-import { detailsSelector } from '../../data/details/selectors';
-import { getPropsToRemoveFractionZeroDigits } from '../../../payment/data/utils';
+import {
+  useIntl,
+  getLocale,
+} from '@edx/frontend-platform/i18n';
 
-const messages = defineMessages({
-  'subscription.checkout.billing.notification': {
-    id: 'subscription.checkout.billing.notification',
-    defaultMessage: 'You will be charged {price} {currency} {trialEnd} then monthly until you cancel your subscription.',
-    description: 'Subscription monthly billing notification for Users that they will be charged every 31 days for this subscription.',
-  },
-  'subscription.checkout.billing.trial.date': {
-    id: 'subscription.checkout.billing.trial.date',
-    defaultMessage: 'on {date},',
-    description: 'Subscription legal trialing helping text.',
-  },
-  'subscription.checkout.billing.resubscribe.date': {
-    id: 'subscription.checkout.billing.resubscribe.date',
-    defaultMessage: 'today,',
-    description: 'Subscription legal resubscribe helping text.',
-  },
-});
+import moment from 'moment-timezone';
+
+import { AppContext } from '@edx/frontend-platform/react';
+import { getPropsToRemoveFractionZeroDigits } from '../../../payment/data/utils';
+import { detailsSelector } from '../../data/details/selectors';
+
+import messages from './messages';
 
 const MonthlyBillingNotification = () => {
   const {
     price, currency, trialEnd, isTrialEligible,
   } = useSelector(detailsSelector);
   const intl = useIntl();
-  const trialDateHelpingText = intl.formatMessage(messages['subscription.checkout.billing.trial.date'], { date: trialEnd });
+  const { authenticatedUser } = useContext(AppContext);
+
+  const userTimezone = (
+    authenticatedUser.timeZone || moment.tz.guess() || 'UTC'
+  );
+  const date = moment(trialEnd);
+  const localizeDate = date.tz(userTimezone);
+
+  // Format the date as "June 28, 2022 (PKT)"
+  const formattedDate = localizeDate.locale(getLocale()).format('MMMM D, YYYY (z)');
+
+  const trialDateHelpingText = intl.formatMessage(messages['subscription.checkout.billing.trial.date'], { date: formattedDate });
   const resubscribeDateHelpingText = intl.formatMessage(messages['subscription.checkout.billing.resubscribe.date'], {});
 
   return (
